+++ deployment/updates/sql/2015_01_13_kuser_add_type_column.sql
@@ -1,2 +1,2 @@
 ALTER TABLE `kuser`
-ADD `type` tinyint(4) DEFAULT '0' AFTER `custom_data`;
+ADD `type` int(11) DEFAULT '0' AFTER `custom_data`;
